+++ src/main/java/com/javarush/model/City.java
@@ -15,7 +15,7 @@ public class City extends BaseEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "city_id", columnDefinition = "smallint")
-    Integer city_id;
+    Short city_id;
     @Column(name = "city", columnDefinition = "varchar(50)")
     String name;
     @ManyToOne(cascade = {CascadeType.MERGE, CascadeType.PERSIST})
